services/news-fetcher/requirements.txt
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
feedparser==6.0.11
aiohttp==3.11.11
python-dateutil==2.9.0
